Moving files into better package hierarchy.
