+++ src/main/scala/com/gu/tagdiffer/flexible/FlexiContent.scala
@@ -1,4 +1,4 @@
-package com.gu.tagdiffer.scaladb
+package com.gu.tagdiffer.flexible
 
 import java.util.NoSuchElementException
