--- src/main/scala/com/gu/tagdiffer/provider/DatabaseProvider.scala
@@ -1,7 +1,0 @@
-package com.gu.tagdiffer.provider
-
-import com.gu.tagdiffer.scaladb.Database
-
-trait DatabaseProvider {
-  def database: Database
-}
