+++ src/main/scala/com/gu/tagdiffer/r2/Database.scala
@@ -1,9 +1,9 @@
-package com.gu.tagdiffer.scaladb
+package com.gu.tagdiffer.r2
 
 import java.sql.{Connection, ResultSet}
 import java.util.Date
 import javax.sql.DataSource
-
+import com.gu.tagdiffer.r2.Query
 import org.joda.time.{DateTime, LocalDate}
 
 import scalaz.Monoid
