+++ src/main/scala/com/gu/tagdiffer/r2/DatabaseComponent.scala
@@ -1,11 +1,11 @@
-package com.gu.tagdiffer.component
+package com.gu.tagdiffer.r2
 
-import javax.sql.DataSource
 import java.util.Properties
-import oracle.jdbc.pool.OracleDataSource
-import com.gu.tagdiffer.scaladb.Database
+import javax.sql.DataSource
+
 import com.gu.tagdiffer.Config
-import com.gu.tagdiffer.provider.DatabaseProvider
+import com.gu.tagdiffer.r2.DatabaseProvider
+import oracle.jdbc.pool.OracleDataSource
 
 trait DatabaseComponent extends DatabaseProvider {
