+++ src/main/scala/com/gu/tagdiffer/r2/DatabaseProvider.scala
@@ -1,0 +1,5 @@
+package com.gu.tagdiffer.r2
+
+trait DatabaseProvider {
+  def database: Database
+}
