+++ src/main/scala/com/gu/tagdiffer/r2/StoredQuery.scala
@@ -1,7 +1,8 @@
-package com.gu.tagdiffer.database
+package com.gu.tagdiffer.r2
+
+import com.gu.tagdiffer.r2.Query
 
 import io.Source
-import com.gu.tagdiffer.scaladb.Query
 import java.util.concurrent.ConcurrentHashMap
 import collection.JavaConversions._
